refactor(role): 역할 정하기 진행 컴포넌트 리팩토링
