+++ frontend/src/pages/RoleMainPage/components/RoleMainRoleEditModalButtonGroup/RoleMainRoleEditModalButtonGroup.tsx
@@ -1,0 +1,35 @@
+import FlexContainer from '../../../../components/FlexContainer/FlexContainer';
+import Button from '../../../../components/Button/Button';
+import { useTheme } from '@emotion/react';
+
+function RoleMainRoleEditModalButtonGroup() {
+  const theme = useTheme();
+  return (
+    <FlexContainer gap="2rem">
+      <Button
+        variant="filled"
+        colorScheme={theme.colors.GRAY_400}
+        width="14rem"
+        padding="1.6rem 3.2rem"
+        borderRadius="1.2rem"
+        fontSize="1.6rem"
+        onClick={close}
+      >
+        취소
+      </Button>
+      <Button
+        variant="filled"
+        colorScheme={theme.colors.YELLOW_200}
+        width="14rem"
+        padding="1.6rem 3.2rem"
+        borderRadius="1.2rem"
+        fontSize="1.6rem"
+        type="submit"
+      >
+        확인
+      </Button>
+    </FlexContainer>
+  );
+}
+
+export default RoleMainRoleEditModalButtonGroup;
