+++ frontend/src/pages/RoleMainPage/components/RoleMainRoleEditModalInputGroup/RoleMainRoleEditModalInputGroup.styles.tsx
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+const StyledBinIcon = styled.img`
+  position: absolute;
+  right: 1.2rem;
+  width: 2rem;
+  cursor: pointer;
+`;
+
+export { StyledBinIcon };
